Split deleting cookies to its separate file
